requirements.txt
opencv-python==4.8.1.78
ultralytics==8.1.0
pytesseract==0.3.10
numpy==1.24.3
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
flask==3.0.0
easyocr==1.7.0
matplotlib==3.8.0
scikit-learn==1.3.2
pandas==2.1.3
pyyaml==6.0.1
tqdm==4.66.1
albumentations==1.3.1
scipy==1.11.4
